requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
arxiv==2.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
mcp==1.9.1
openai==1.82.0
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.2
